requirements.txt
absl-py==0.11.0
argon2-cffi==20.1.0
astunparse==1.6.3
async-generator==1.10
async-timeout==3.0.1
attrs==20.3.0
backcall==0.2.0
bleach==3.2.1
blinker==1.4
Brotli==1.0.9
brotlipy==0.7.0
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
cmapy==0.6.6
cmdstanpy==0.9.5
convertdate==2.3.0
cycler==0.10.0
Cython==0.29.21
dash==1.19.0
dash-core-components==1.15.0
dash-html-components==1.1.2
dash-renderer==1.9.0
dash-table==4.11.2
dash-table-experiments==0.6.0
dataclasses==0.6
decorator==4.4.2
defusedxml==0.6.0
docopt==0.6.2
entrypoints==0.3
ephem==3.7.7.1
fbprophet==0.7.1
Flask==1.1.2
Flask-Compress==1.8.0
flatbuffers==1.12
future==0.18.2
gast==0.3.3
google-auth==1.23.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grip==4.5.2
grpcio==1.32.0
h5py==2.10.0
hijri-converter==2.1.1
holidays==0.10.5.2
idna==2.10
imageio==2.9.0
ipykernel==5.3.4
ipython==7.19.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.2
joblib==0.17.0
json5==0.9.5
jsonschema==3.2.0
jupyter-client==6.1.7
jupyter-core==4.6.3
jupyterlab==2.2.5
jupyterlab-pygments==0.1.2
jupyterlab-server==1.2.0
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
korean-lunar-calendar==0.2.1
LunarCalendar==0.0.9
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.3.3
mistune==0.8.4
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.7
nest-asyncio==1.4.3
networkx==2.5
notebook==6.1.5
numpy==1.19.4
oauthlib==3.1.0
olefile==0.46
opencv-contrib-python==4.4.0.44
opencv-python==4.4.0.44
opt-einsum==3.3.0
ordered-set==4.0.2
packaging==20.4
pandas==1.2.1
pandocfilters==1.4.2
parso==0.7.0
path-and-address==2.0.1
pathlib==1.0.1
patsy==0.5.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.1
plotly==4.14.3
prometheus-client==0.8.0
prompt-toolkit==3.0.8
protobuf==3.13.0
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.2
pycparser==2.20
Pygments==2.7.2
PyJWT==1.7.1
PyMeeus==0.3.7
pyparsing==2.4.7
pyrsistent==0.17.3
PySocks==1.7.1
pystan==2.19.1.1
python-dateutil==2.8.1
pytz==2020.1
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==19.0.2
requests==2.25.0
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.6
scikit-image==0.17.2
scikit-learn==0.23.2
scipy==1.5.4
Send2Trash==1.5.0
setuptools-git==1.2
sip==5.0.0
six==1.15.0
sklearn==0.0
split-folders==0.4.3
statsmodels==0.12.2
tensorboard==2.4.1
tensorboard-plugin-wit==1.7.0
tensorflow==2.4.1
tensorflow-estimator==2.4.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
threadpoolctl==2.1.0
tifffile==2020.10.1
toml==0.10.2
torch==1.7.1
torchmeta==1.6.0
torchsummary==1.5.1
torchvision==0.8.2
tornado==6.0.4
tqdm==4.52.0
traitlets==5.0.5
typing-extensions==3.7.4.3
urllib3==1.26.2
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0
